2.1.0 Improve build | Add ESM support | Code refactoring | Other improvements
