+++ src/index.d.ts
@@ -7,7 +7,7 @@ declare module "passfather" {
     symbols?: boolean | undefined;
     length?: number | undefined;
     ranges?: Array<number[]> | undefined;
-    prng?: 'default' | 'Alea' | 'KISS07' | 'Kybos' | 'LFib' | 'LFIB4' | 'MRG32k3a' | 'Xorshift03';
+    prng?: 'default' | 'Alea' | 'KISS07' | 'Kybos' | 'LFib' | 'LFIB4' | 'MRG32k3a' | 'Xorshift03' | undefined;
     seed?: Array<string | number> | undefined;
   }
