+++ src/seed.js
@@ -1,11 +1,11 @@
-const { compact, isBrowser } = require('./utils');
-const os = isBrowser() ? null : eval('require("os")');
+const { compact, hasWindow } = require('./utils');
+const os = hasWindow ? {} : eval(`require('os')`);
 
 /**
  * Default seed for prng
  * @const
  */
-const DEFAULT_NODE_SEED = !isBrowser() ? compact([].concat(
+const DEFAULT_NODE_SEED = !hasWindow ? compact([].concat(
   Date.now(),
   (
     process.memoryUsage
@@ -31,7 +31,7 @@ const DEFAULT_NODE_SEED = !isBrowser() ? compact([].concat(
  * Default seed for prng
  * @const
  */
-const DEFAULT_BROWSER_SEED = isBrowser() ? compact([].concat(
+const DEFAULT_BROWSER_SEED = hasWindow ? compact([].concat(
   Date.now(),
   (
     performance && performance.memory
