+++ test/package.json
@@ -1,4 +1,5 @@
 {
+  "private": true,
   "name": "passfather-test",
   "version": "1.0.0",
   "description": "Testing passfather",
